更新部分文件

+++ app/src/main/java/com/ezlinker/app/modules/role/model/Role.java
@@ -8,7 +8,7 @@ import lombok.experimental.Accessors;
 
 /**
  * <p>
- * 
+ *
  * </p>
  *
  * @author wangwenhai
@@ -26,7 +26,7 @@ public class Role extends XEntity {
 
     private String name;
 
-    private Integer parent;
+    private Long parent;
 
     private String description;
